Added specific README for OS X
The goal would be to remove as much dependencies as possible.

I am hoping to only require opencv and boost. I will use OpenCv's
highgui and Cocoa's UI whenever possible and remove the dependencies to
vxl, gtkmm and cairomm because those take up a lot of space on a user's
machine.